apps/minecraft-api/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Pillow==11.1.0
whitenoise==6.9.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-colorfield==0.11.0
django-redis==5.4.0
httpx==0.28.1
groq==0.22.0
